requirements-prod.txt
# Production Requirements
# MLOps & Experiment Tracking
mlflow==2.9.2
psycopg2-binary==2.9.9

# API Server
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# Caching & Performance
redis==5.0.1
hiredis==2.3.2

# Real-time Streaming
kafka-python==2.0.2
websockets==12.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Performance Optimization
onnx==1.15.0
onnxruntime==1.16.3
tensorflow-model-optimization==0.7.5

# Distributed Training
horovod==0.28.1

# Additional utilities
python-dotenv==1.0.0
aiocache==0.12.2
httpx==0.26.0
